Update Sentinel_linear_search.cpp
Program to implement Sentinel Linear Search.
Sentinel Linear search is a type of linear search where the element to be searched is placed in the last position and then all the indices are checked for the presence of the element without checking for the index out of bound case. (source for definition GeeksforGeeks).
Its a better approach as compared to normal linear search as number to operations greatly reduced from number 2*n+1 to n+2 in  worst case.